Create DEs and Volcano plot
The gene expression matrix of the control group and the test group samples was imported, and the up-regulated genes and down-regulated genes were screened according to the threshold of log2FoldChange and p value, and the corresponding volcano plot was drawn. The gene expression matrix of all samples of the up-regulated genes and down-regulated genes after screening was derived.